cryptography/requirements.txt
pycryptodomex
pycryptodome
loguru
boto3
pyspark
findspark
mysql-connector-python
python-resources